fix max2881 info mess
