+++ firmware/application/hw/max2831.cpp
@@ -38,19 +38,20 @@ using namespace hackrf::one;
 #include <algorithm>
 #include <cstring>
 
-// Global debug tracking for MAX2831
-struct max2831_debug_t {
-    uint32_t requested_freq_mhz;
-    uint32_t calculated_n;
-    uint32_t calculated_frac;
-    bool set_frequency_called;
-    bool frequency_valid;
-};
-extern "C" max2831_debug_t max2831_debug_info ;
-
 namespace max2831 {
 
 using namespace max283x;
+
+static MAX2831Info max2831_info = {0, 0, 0, false, false};
+
+MAX2831Info get_max2831_info() {
+    return {
+        max2831_info.requested_freq_mhz,
+        max2831_info.calculated_n,
+        max2831_info.calculated_frac,
+        max2831_info.set_frequency_called,
+        max2831_info.frequency_valid};
+}
 
 /*
  * MAX2831 uses 9-bit SPI transfers.
@@ -385,13 +386,13 @@ bool MAX2831::set_frequency(const rf::Frequency lo_frequency) {
     bool valid = (lo_frequency >= 2300000000LL && lo_frequency <= 2600000000LL);
 
     // TRACK REQUEST IMMEDIATELY
-    max2831_debug_info.requested_freq_mhz = lo_frequency / 1000000;
-    max2831_debug_info.set_frequency_called = true;
-    max2831_debug_info.frequency_valid = valid;
+    max2831_info.requested_freq_mhz = lo_frequency / 1000000;
+    max2831_info.set_frequency_called = true;
+    max2831_info.frequency_valid = valid;
 
     if (!valid) {
-        max2831_debug_info.calculated_n = 0;
-        max2831_debug_info.calculated_frac = 0;
+        max2831_info.calculated_n = 0;
+        max2831_info.calculated_frac = 0;
         return false;
     }
 
@@ -414,8 +415,8 @@ bool MAX2831::set_frequency(const rf::Frequency lo_frequency) {
     }
 
     // TRACK CALCULATED VALUES
-    max2831_debug_info.calculated_n = div_int;
-    max2831_debug_info.calculated_frac = div_frac;
+    max2831_info.calculated_n = div_int;
+    max2831_info.calculated_frac = div_frac;
 
     /* Write order matters - matches GSG reference */
     /* REG 3: SYN_INT (bits 7:0) and SYN_FRAC_LO (bits 13:8) */
